callcenter-ai/backend/requirements/base.txt
# requirements/base.txt
# Core Django and essential packages for all environments
Django==5.2.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-tenants==3.8.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-phonenumber-field==8.1.0
django-simple-history==3.8.0
django-import-export==4.3.7
python-decouple==3.8
python-dotenv==1.1.0

# Database adapters
psycopg2-binary==2.9.10
mysqlclient==2.2.7

# Task queue
celery==5.5.2
redis==6.1.0
amqp==5.3.1
billiard==4.2.1
kombu==5.5.3
vine==5.1.0

# Image processing
Pillow==10.0.0

# API documentation
drf-yasg==1.21.10

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
PyYAML==6.0.2
six==1.17.0
packaging==25.0
uritemplate==4.1.1
inflection==0.5.1
PyJWT==2.9.0

# Django core dependencies
asgiref==3.8.1
sqlparse==0.5.3

# Click (Celery dependency)
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0

# Tablib (django-import-export dependency)
tablib==3.8.0

jsonschema